pyproject.toml
[project]
authors = [
    { name = "Fabrice Normandin", email = "fabrice.normandin@gmail.com" },
]
name = "simple-parsing"
dynamic = ["version"]
description = "A small utility to simplify and clean up argument parsing scripts."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["docstring-parser~=0.15", "typing-extensions>=4.5.0"]
license = { file = "LICENSE" }

[project.optional-dependencies]
yaml = ["pyyaml>=6.0.2"]
toml = ["tomli>=2.2.1", "tomli-w>=1.0.0"]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["I", "UP"]
# TODO: Fix the imports in __init__.py files and enable these:
# select = ["E4", "E7", "E9", "I", "UP"]

[tool.docformatter]
in-place = true
wrap-summaries = 99
wrap-descriptions = 99

[tool.codespell]
skip = ["logs/**", "data/**"]

[dependency-groups]
dev = [
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pytest>=8.3.4",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.0.0",
    "pytest-regressions>=2.7.0",
    "pytest-xdist>=3.6.1",
]
